+++ frames/popup/pop_new_type.py
@@ -1,5 +1,5 @@
 from tkinter import Toplevel, Button, Label, Entry, OptionMenu, StringVar, E, W
-from services import data_services
+from services import db_services
 from services import gui_services
 
 class PopNewType(Toplevel):
