+++ main.py
@@ -1,5 +1,5 @@
 from frames.window_manager import WindowManager as pockets
-from services import data_services
+from services import db_services
 
 
 def run(app):
@@ -7,4 +7,4 @@ def run(app):
 
 
 if __name__ == "__main__":
-    run(pockets(data_services.get_database_connection()))
+    run(pockets(db_services.get_database_connection()))
